pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurobm"
version = "0.1.0"
description = "Educational/Research Platform for Modeling Human Brain Dynamics with Boltzmann Machines"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "NeuroBM Contributors", email = "neurobm@example.com"}
]
maintainers = [
    {name = "NeuroBM Contributors", email = "neurobm@example.com"}
]
keywords = [
    "boltzmann-machine",
    "deep-learning",
    "cognitive-modeling",
    "neuroscience",
    "interpretability",
    "research",
    "education"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Education",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
requires-python = ">=3.12"
dependencies = [
    # Core ML/Scientific Computing
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",

    # Data handling and visualization
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "ipywidgets>=8.0.0",

    # Configuration and utilities
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "typer>=0.9.0",

    # Jupyter and notebook support
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "nbformat>=5.9.0",

    # Validation and schema
    "pydantic>=2.0.0",
    "jsonschema>=4.17.0",

    # Logging and monitoring
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",

    # Statistical analysis
    "statsmodels>=0.14.0",
    "pingouin>=0.5.3",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-mock>=3.11.0",

    # Code quality
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",

    # Documentation
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",

    # Profiling and debugging
    "line-profiler>=4.1.0",
    "memory-profiler>=0.61.0",
    "py-spy>=0.3.14",
]

docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "nbsphinx>=0.9.0",
]

all = [
    "neurobm[dev,docs]"
]

[project.urls]
Homepage = "https://github.com/neurobm/neurobm"
Documentation = "https://neurobm.readthedocs.io"
Repository = "https://github.com/neurobm/neurobm"
"Bug Tracker" = "https://github.com/neurobm/neurobm/issues"

[project.scripts]
neurobm = "neurobm.cli:main"
neurobm-train = "scripts.train:main"
neurobm-sample = "scripts.sample:main"
neurobm-eval = "scripts.eval_ais:main"
neurobm-scaffold = "scripts.neurobm_scaffold:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["neurobm*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
neurobm = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["neurobm"]
known_third_party = ["torch", "numpy", "pandas", "matplotlib", "seaborn", "sklearn"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "wandb.*",
    "tensorboard.*",
    "plotly.*",
    "seaborn.*",
    "sklearn.*",
    "scipy.*",
    "pingouin.*",
    "statsmodels.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
]

[tool.coverage.run]
source = ["neurobm"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/.*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]